refactor create character forms
