+++ apps/game-master-client/app/routes/_app.games.$gameId.characters.new/route.tsx
@@ -1,7 +1,7 @@
 import type { ActionFunctionArgs } from "@remix-run/node";
 import { useActionData, useParams } from "@remix-run/react";
 import { useState } from "react";
-import { CreateCharacterForm } from "~/components/forms/create-character";
+import { CreateCharacterForm } from "~/components/forms/character-forms";
 import { Text } from "~/components/ui/typeography";
 import { createCharacterAction } from "~/queries/server/create-character.server";
